Move fetch contact logic inside AddContact screeen
